+++ skysail.server.ext.ramlconsole.webresource/src/io/skysail/server/ext/ramlconsole/webresource/RamlConsoleConstants.java
@@ -1,0 +1,7 @@
+package io.skysail.server.ext.ramlconsole.webresource;
+
+public interface RamlConsoleConstants {
+    //String WEB_RESOURCE_SHORTNAME = "ramlconsole";
+    String WEB_RESOURCE_NAME = "/mulsesoft/raml-console";
+    String WEB_RESOURCE_VERSION = "2.0.5";
+}