+++ skysail.server/src/de/twenty11/skysail/server/core/restlet/AnyOf.java
@@ -1,5 +1,0 @@
-package de.twenty11.skysail.server.core.restlet;
-
-public class AnyOf {
-
-}