+++ skysail.server/src/de/twenty11/skysail/server/core/restlet/RoleAuthorizerFactory.java
@@ -1,9 +1,0 @@
-package de.twenty11.skysail.server.core.restlet;
-
-import org.restlet.security.Authorizer;
-
-public interface RoleAuthorizerFactory {
-
-    Authorizer create(String roleName);
-
-}