+++ skysail.server/src/io/skysail/server/services/DecryptorService.java
@@ -1,10 +1,10 @@
-package io.skysail.server.services;
-
-import org.osgi.annotation.versioning.ProviderType;
-
-@ProviderType
-public interface DecryptorService {
-
-    String decryptText(String text, String password);
-
-}
+//package io.skysail.server.services;
+//
+//import org.osgi.annotation.versioning.ProviderType;
+//
+//@ProviderType
+//public interface DecryptorService {
+//
+//    String decryptText(String text, String password);
+//
+//}