+++ skysail.server/etc/changelog.md
@@ -6,4 +6,7 @@ new in
  * new method SkysailApplication#createStaticDirectory() to add static content from inside the bundle
  * implemented: CORS config (https://github.com/evandor/skysail/issues/1)
  * implemented: removed polymer in bootstrap theme (https://github.com/evandor/skysail/issues/2)
+ * removed AddApiVersionHeaderFilter
+ * OptionalEncryptionFilter
+ * removed AtomicReferences for OSGi services